Extracted model application out of PercolatorFdr into PercolatorScorer
* Full-population scoring, the projection-native in-place path and the streaming
  first pass, with the per-row primitives they share
* PercolatorFdr.cs 3,933 -> 2,845 lines; 1,087 moved lines verified verbatim
* ScoreWithFoldModel is internal, not private: training calls it to score its
  held-out folds, so the training-onto-scoring dependency is explicit
* Four q-value helpers widened to internal, reflecting the pre-existing
  scoring/FDR conflation that the FDR extraction will separate

See ai/todos/active/TODO-20260727_osprey_percolatorfdr_decomposition.md
