+++ pwiz_tools/Osprey/Osprey.FDR/FdrProjectionOutput.cs
@@ -101,7 +101,7 @@ namespace pwiz.Osprey.FDR
     /// FdrProjection struct-shrink S0 / increment C1). The lean
     /// <see cref="FdrProjection"/> no longer carries the six q-value outputs; the
     /// streaming write-back path
-    /// (<c>PercolatorFdr.ScoreProjectionAndComputeFdrInPlace</c>) instead calls
+    /// (<c>PercolatorScorer.ScoreProjectionAndComputeFdrInPlace</c>) instead calls
     /// <see cref="Accept"/> per row, in the existing nested (file, row) order, handing
     /// each row's freshly computed <see cref="FdrQValues"/> + <see cref="FdrProjection.Score"/>
     /// to a caller-supplied sink. Both passes stream every row straight to the per-file
